+++ src/hooks.ts
@@ -1,7 +1,7 @@
 import { MODULE_ID } from "./constants";
-import { ChatNotifier } from "./features/chat/chat-notifier";
-import { CHAT_SETTINGS } from "./features/chat/constants";
-import { ChatSettingsMenu } from "./features/chat/settings-menu";
+import { injectConditionFilter } from "./features/misc/condition-filter";
+import { MISC_SETTINGS } from "./features/misc/constants";
+import { MiscSettingsMenu } from "./features/misc/settings-menu";
 import { NAMEPLATE_SETTINGS, NAMEPLATE_TOKEN_FLAGS } from "./features/nameplate/constants";
 import { NameplateFitter } from "./features/nameplate/nameplate-fitter";
 import { NameplateSettingsMenu } from "./features/nameplate/settings-menu";
@@ -16,7 +16,7 @@ import { registerModuleSettings } from "./settings";
 
 Hooks.once("init", () => {
     registerModuleSettings(
-        [NAMEPLATE_SETTINGS, TOOLTIP_SETTINGS, SPELL_FADE_SETTINGS, CHAT_SETTINGS],
+        [NAMEPLATE_SETTINGS, TOOLTIP_SETTINGS, SPELL_FADE_SETTINGS, MISC_SETTINGS],
         {
             [NAMEPLATE_SETTINGS.ENABLED.key]: () => NameplateFitter.refreshAll(),
             [NAMEPLATE_SETTINGS.MIN_FONT_SIZE.key]: () => NameplateFitter.refreshAll(),
@@ -59,12 +59,12 @@ Hooks.once("init", () => {
         restricted: false
     });
 
-    game.settings.registerMenu(MODULE_ID, "chatSettingsMenu", {
-        name: "imaginaryfriend-toolkit.Menus.chat.Name",
-        label: "imaginaryfriend-toolkit.Menus.chat.Label",
-        hint: "imaginaryfriend-toolkit.Menus.chat.Hint",
-        icon: "fa-solid fa-bell",
-        type: ChatSettingsMenu,
+    game.settings.registerMenu(MODULE_ID, "miscSettingsMenu", {
+        name: "imaginaryfriend-toolkit.Menus.misc.Name",
+        label: "imaginaryfriend-toolkit.Menus.misc.Label",
+        hint: "imaginaryfriend-toolkit.Menus.misc.Hint",
+        icon: "fa-solid fa-ellipsis",
+        type: MiscSettingsMenu,
         restricted: false
     });
 
@@ -119,4 +119,4 @@ Hooks.on("controlToken", (token: Token, controlled: boolean) => {
 Hooks.on("renderTokenConfig", (app, htmlElement: HTMLElement) => injectTokenConfigField(app, htmlElement));
 Hooks.on("renderPrototypeTokenConfig", (app, htmlElement: HTMLElement) => injectTokenConfigField(app, htmlElement));
 
-Hooks.on("createChatMessage", (message: ChatMessage) => ChatNotifier.onCreateMessage(message));
+Hooks.on("renderTokenHUD", (app, htmlElement: HTMLElement) => injectConditionFilter(app, htmlElement));